+++ lib/src/tt_response.dart
@@ -1,0 +1,109 @@
+/// Platform channel request/response key constants.
+class TTResponse {
+  TTResponse._();
+
+  static const String command = "command";
+  static const String data = "data";
+  static const String resultState = "resultState";
+  static const String errorCode = "errorCode";
+  static const String errorMessage = "errorMessage";
+  static const String state = "state";
+  static const String status = "status";
+  static const String wifiList = "wifiList";
+  static const String finished = "finished";
+
+  static const String lockName = "lockName";
+  static const String lockMac = "lockMac";
+  static const String isInited = "isInited";
+  static const String isAllowUnlock = "isAllowUnlock";
+  static const String isDfuMode = "isDfuMode";
+  static const String electricQuantity = "electricQuantity";
+  static const String lockVersion = "lockVersion";
+  static const String lockSwitchState = "lockSwitchState";
+  static const String rssi = "rssi";
+  static const String oneMeterRssi = "oneMeterRssi";
+  static const String timestamp = "timestamp";
+  static const String specialValue = "specialValue";
+  static const String lockData = "lockData";
+  static const String lockConfig = "lockConfig";
+  static const String scanState = "scanState";
+  static const String passcodeInfo = "passcodeInfo";
+  static const String controlAction = "controlAction";
+  static const String lockTime = "lockTime";
+  static const String uniqueId = "uniqueId";
+  static const String passcode = "passcode";
+  static const String startDate = "startDate";
+  static const String endDate = "endDate";
+  static const String passcodeOrigin = "passcodeOrigin";
+  static const String passcodeNew = "passcodeNew";
+  static const String cardNumber = "cardNumber";
+  static const String fingerprintNumber = "fingerprintNumber";
+  static const String adminPasscode = "adminPasscode";
+  static const String erasePasscode = "erasePasscode";
+  static const String totalCount = "totalCount";
+  static const String currentCount = "currentCount";
+  static const String records = "records";
+  static const String maxTime = "maxTime";
+  static const String cycleJsonList = "cycleJsonList";
+  static const String faceFeatureData = "faceFeatureData";
+  static const String faceNumber = "faceNumber";
+
+  static const String minTime = "minTime";
+  static const String currentTime = "currentTime";
+  static const String isOn = "isOn";
+  static const String passageModeType = "passageModeType";
+  static const String weekly = "weekly";
+  static const String monthly = "monthly";
+
+  static const String direction = "direction";
+
+  static const String isSupport = "isSupport";
+  static const String supportFunction = "supportFunction";
+
+  static const String nbAwakeModes = "nbAwakeModes";
+  static const String nbAwakeTimeList = "nbAwakeTimeList";
+  static const String minutes = "minutes";
+  static const String type = "type";
+  static const String hotelInfo = "hotelInfo";
+  static const String buildingNumber = "buildingNumber";
+  static const String floorNumber = "floorNumber";
+  static const String sector = "sector";
+
+  static const String passcodeListString = "passcodeListString";
+  static const String cardListString = "cardListString";
+  static const String fingerprintListString = "fingerprintListString";
+
+  static const String nbServerAddress = "nbServerAddress";
+  static const String nbServerPort = "nbServerPort";
+
+  static const String modelNum = "modelNum";
+  static const String hardwareRevision = "hardwareRevision";
+  static const String firmwareRevision = "firmwareRevision";
+  static const String nbNodeId = "nbNodeId";
+  static const String nbOperator = "nbOperator";
+  static const String nbCardNumber = "nbCardNumber";
+  static const String nbRssi = "nbRssi";
+
+  static const String addGatewayJsonStr = "addGatewayJsonStr";
+  static const String ip = "ip";
+  static const String port = "port";
+
+  static const String ipSettingJsonStr = "ipSettingJsonStr";
+  static const String wifiName = "wifiName";
+  static const String wifiPassword = "wifiPassword";
+
+  static const String mac = "mac";
+
+  static const String remoteAccessory = "remoteAccessory";
+
+  static const String soundVolumeType = "soundVolumeType";
+
+  static const String updateDate = "updateDate";
+  static const String alertTime = "alertTime";
+  static const String wirelessKeypadFeatureValue = "wirelessKeypadFeatureValue";
+  static const String resetCode = "resetCode";
+
+  static const String sensitivityValue = "sensitivityValue";
+
+  static const String videoModuleSerialNumber = "videoModuleSerialNumber";
+}