+++ lib/ttdoorSensor.dart
@@ -1,16 +1,14 @@
+import 'package:ttlock_premise_flutter/src/ttlock_channel.dart';
+import 'package:ttlock_premise_flutter/src/ttdoorsensor_commands.dart';
 import 'package:ttlock_premise_flutter/ttlock.dart';
 
 class TTDoorSensor {
-  static const String COMMAND_START_SCAN_DOOR_SENSOR = "doorSensorStartScan";
-  static const String COMMAND_STOP_SCAN_DOOR_SENSOR = "doorSensorStopScan";
-  static const String COMMAND_INIT_DOOR_SENSOR = "doorSensorInit";
-
   static void startScan(TTRemoteAccessoryScanCallback scanCallback) {
-    TTLock.invoke(COMMAND_START_SCAN_DOOR_SENSOR, null, scanCallback);
+    TTLockChannel.invoke(TTDoorSensorCommands.COMMAND_START_SCAN_DOOR_SENSOR, null, scanCallback);
   }
 
   static void stopScan() {
-    TTLock.invoke(COMMAND_STOP_SCAN_DOOR_SENSOR, null, null);
+    TTLockChannel.invoke(TTDoorSensorCommands.COMMAND_STOP_SCAN_DOOR_SENSOR, null, null);
   }
 
   static void init(
@@ -22,7 +20,7 @@ class TTDoorSensor {
     Map map = Map();
     map[TTResponse.mac] = mac;
     map[TTResponse.lockData] = lockData;
-    TTLock.invoke(COMMAND_INIT_DOOR_SENSOR, map, callback,
+    TTLockChannel.invoke(TTDoorSensorCommands.COMMAND_INIT_DOOR_SENSOR, map, callback,
         fail: failedCallback);
   }
 }